advance-rag/pyproject.toml
[project]
name = "advance-rag"
version = "0.1.0"
description = "RAG pipeline service for b-knowledge, extracted from RAGFlow"
requires-python = ">=3.10"
dependencies = [
    # Environment
    "python-dotenv>=1.0.0",

    # Logging
    "loguru>=0.7.0",

    # Database (Peewee + PostgreSQL)
    "peewee>=3.17.0",
    "psycopg2-binary>=2.9.9",

    # Core infrastructure
    "redis>=5.0.0",
    "minio>=7.2.0",
    "opensearch-py>=2.0.0",
    "ruamel.yaml>=0.18.0",
    "strenum>=0.4.15",
    "filelock>=3.13.0",
    "xxhash>=3.4.0",
    "itsdangerous>=2.1.0",
    "pymysql>=1.1.0",
    "pyobvector>=0.2.0",

    # Cloud storage
    "azure-storage-blob>=12.0.0",
    "azure-storage-file-datalake>=12.0.0",
    "azure-identity>=1.0.0",
    "google-cloud-storage>=2.0.0",
    "boto3>=1.28.0",
    "opendal>=0.45.0",

    # NLP / ML
    "transformers>=4.35.0",
    "tokenizers>=0.15.0",
    "numpy>=1.24.0",
    "onnxruntime>=1.16.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "graspologic>=3.4.0",
    "roman-numbers>=1.0.0",
    "word2number>=1.1",
    "cn2an>=0.5.0",
    "hanziconv>=0.3.0",
    "beartype>=0.18.0",
    "tiktoken>=0.5.0",
    "networkx>=3.0",
    "pandas>=2.0.0",
    "nltk>=3.8.0",

    # Document parsing
    "pdfplumber>=0.10.0",
    "pdfminer.six>=20231228",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.5.0",
    "mammoth>=1.6.0",
    "tika>=2.6.0",
    "trafilatura>=1.6.0",
    "markdownify>=0.11.0",
    "msglite>=0.30.0",
    "olefile>=0.47",
    "markdown-to-json>=2.1.0",

    # OCR / Vision
    "opencv-python>=4.8.0",
    "shapely>=2.0.0",
    "pyclipper>=1.3.0",
    "editdistance>=0.6.0",
    "huggingface_hub>=0.20.0",

    # LLM providers (Ollama, OpenAI, Azure, Google Vertex)
    "openai>=1.0.0",
    "langfuse>=2.0.0",

    # Additional runtime dependencies
    "infinity-sdk>=0.6.0",
    "valkey>=6.0.0",
    "json_repair>=0.28.0",
    "pypdf>=4.0.0",
    "litellm>=1.0.0",
    "ollama>=0.3.0",

    # Utilities
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
    "argon2-cffi>=23.0.0",
    "demjson3>=3.0.0",
    "xpinyin>=0.7.0",
    "retry>=0.9.0",
    "websocket-client>=1.6.0",
    "exceptiongroup>=1.0.0",
    "werkzeug>=3.0.0",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "jinja2>=3.1.0",
    "six>=1.16.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["common*", "db*", "deepdoc*", "rag*", "memory*"]
